+++ control-plane/api/v1alpha1/servicedefaults_types.go
@@ -1,7 +1,6 @@
 package v1alpha1
 
 import (
-	"encoding/json"
 	"fmt"
 	"net"
 	"strings"
@@ -189,18 +188,6 @@ type ServiceDefaultsDestination struct {
 	// Destination.
 	Port uint32 `json:"port,omitempty"`
 }
-
-// EnvoyExtension has configuration for an extension that patches Envoy resources.
-type EnvoyExtension struct {
-	Name     string `json:"name,omitempty"`
-	Required bool   `json:"required,omitempty"`
-	// +kubebuilder:validation:Type=object
-	// +kubebuilder:validation:Schemaless
-	// +kubebuilder:pruning:PreserveUnknownFields
-	Arguments json.RawMessage `json:"arguments,omitempty"`
-}
-
-type EnvoyExtensions []EnvoyExtension
 
 func (in *ServiceDefaults) ConsulKind() string {
 	return capi.ServiceDefaults
@@ -517,50 +504,4 @@ func (in *ServiceDefaults) MatchesConsul(candidate capi.ConfigEntry) bool {
 
 func (in *ServiceDefaults) ConsulGlobalResource() bool {
 	return false
-}
-
-func (in EnvoyExtensions) toConsul() []capi.EnvoyExtension {
-	if in == nil {
-		return nil
-	}
-
-	outConfig := make([]capi.EnvoyExtension, 0)
-
-	for _, e := range in {
-		consulExtension := capi.EnvoyExtension{
-			Name:     e.Name,
-			Required: e.Required,
-		}
-
-		// We already validate that arguments is present
-		var args map[string]interface{}
-		_ = json.Unmarshal(e.Arguments, &args)
-		consulExtension.Arguments = args
-		outConfig = append(outConfig, consulExtension)
-	}
-
-	return outConfig
-}
-
-func (in EnvoyExtensions) validate(path *field.Path) field.ErrorList {
-	if len(in) == 0 {
-		return nil
-	}
-
-	var errs field.ErrorList
-	for i, e := range in {
-		if err := e.validate(path.Child("envoyExtension").Index(i)); err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	return errs
-}
-
-func (in EnvoyExtension) validate(path *field.Path) *field.Error {
-	if in.Arguments == nil {
-		err := field.Required(path.Child("arguments"), "arguments must be defined")
-		return err
-	}
-	return nil
 }
